+++ src/provider/client/request.ts
@@ -230,7 +230,7 @@ function extractRateLimitHeaders(
   // Qualys docs indicate that when there is no `x-ratelimit-limit`, and there
   // is a `x-concurrency-limit-limit`, then we've hit the concurrency limit and
   // need to wait.
-  if (!limit && concurrency) {
+  if ((!limit || limit === '0') && concurrency) {
     toWaitSeconds = rateLimitConfig.concurrencyDelay / 1000;
   }